refactor!: remove async wherever not necessary

+++ lua/orgmode/attach/fileops.lua
@@ -16,21 +16,6 @@ local M = {}
 -- libuv functions ported to use OrgPromise
 --]]
 
----Like `vim.uv.fs_readlink`, but returns a promise.
----@param path string
----@return OrgPromise<string> path
-function M.readlink(path)
-  return Promise.new(function(resolve, reject)
-    uv.fs_readlink(path, function(err, link)
-      if link then
-        resolve(link)
-      else
-        reject(err)
-      end
-    end)
-  end)
-end
-
 ---Like `vim.uv.fs_rename`, but returns a promise.
 ---@param path string
 ---@param new_path string
@@ -100,52 +85,6 @@ end
 -- convenient.
 --]]
 
----Get file status.
----* `preserve_symlinks`: if true, do not resolve symlinks, like
----  `vim.uv.fs_lstat`. The default is to resolve symlinks, like
----  `vim.uv.fs_stat`.
----@param path string
----@param opts? {preserve_symlinks: boolean?}
----@return OrgPromise<uv.fs_stat.result> stat
-function M.stat(path, opts)
-  return Promise.new(function(resolve, reject)
-    local preserve_symlinks = opts and opts.preserve_symlinks
-    local fstat = preserve_symlinks and uv.fs_lstat or uv.fs_stat
-    fstat(path, function(err, stat)
-      if stat then
-        resolve(stat)
-      else
-        reject(err)
-      end
-    end)
-  end)
-end
-
----Like `stat`, but resolve to nil if the file does not exist.
----This wrapper simply catches `ENOENT` and resolves as successful nil in its
----stead.
----* `preserve_symlinks`: if true, do not resolve symlinks, like
----  `vim.uv.fs_lstat`. The default is to resolve symlinks, like
----  `vim.uv.fs_stat`.
----@param path string
----@param opts? {preserve_symlinks: boolean?}
----@return OrgPromise<uv.fs_stat.result | nil> stat
-function M.stat_or_nil(path, opts)
-  return Promise.new(function(resolve, reject)
-    local preserve_symlinks = opts and opts.preserve_symlinks
-    local fstat = preserve_symlinks and uv.fs_lstat or uv.fs_stat
-    fstat(path, function(err, stat)
-      if stat then
-        resolve(stat)
-      elseif err and vim.startswith(err, 'ENOENT:') then
-        resolve(nil)
-      else
-        reject(err)
-      end
-    end)
-  end)
-end
-
 ---Like `vim.uv.fs_scandir`, but return an iterator.
 ---Each iteration yields a pair `(name, filetype)` as would be returned by
 ---`vim.uv.fs_scandir_next`.
@@ -210,16 +149,13 @@ function M.make_dir(path, opts)
   return Promise.new(function(resolve, reject)
     uv.fs_mkdir(path, mode, function(err)
       if not err then
-        return resolve(false)
+        return resolve(true)
       elseif vim.startswith(err, 'EEXIST:') and exist_ok then
-        M.is_dir(path):next(function(is_dir)
-          if is_dir then
-            resolve(false)
-          else
-            error(err)
-          end
-          return nil
-        end)
+        if M.is_dir(path) then
+          resolve(false)
+        else
+          error(err)
+        end
         return
       elseif vim.startswith(err, 'ENOENT:') and parents then
         -- Remove trailing slashes.
@@ -230,13 +166,12 @@ function M.make_dir(path, opts)
         if parent == path then
           return reject(err)
         end
-        M.make_dir(parent, { mode = mode, parents = true, exist_ok = false })
-            :next(function()
-              return M.make_dir(path, {
-                mode = mode, parents = false, exist_ok = false })
-            end)
-        ---@diagnostic disable-next-line
-            :next(resolve, reject)
+        M.make_dir(parent, { mode = mode, parents = true, exist_ok = false }):next(function()
+          return M
+              .make_dir(path, { mode = mode, parents = false, exist_ok = false })
+              ---@diagnostic disable-next-line: redundant-parameter
+              :next(resolve, reject)
+        end)
       else
         return reject(err)
       end
@@ -248,66 +183,24 @@ end
 -- Additional functionality that builds upon libuv.
 --]]
 
----Resolve to true if the given file exists, to false otherwise.
----* `preserve_symlinks`: if true, do not resolve symlinks. A broken symlink
----  would still appear as an existing file in this case.
+---Return true if the path points at a directory.
+---This simply uses `fs_scandir()` (as it is slightly faster than `fs_stat()`),
+---so it always resolves symlinks.
 ---@param path string
----@param opts? {preserve_symlinks: boolean?}
----@return OrgPromise<boolean> result
-function M.exists(path, opts)
-  return M.stat_or_nil(path, opts):next(function(stat)
-    return stat and true or false
-  end)
+---@return boolean result
+function M.is_dir(path)
+  local dir, errmsg, err = uv.fs_scandir(path)
+  assert(err == 'ENOTDIR' or err == 'ENOENT', errmsg)
+  return dir and true or false
 end
 
----Resolve to true if the path points at a regular file.
----* `preserve_symlinks`: if true, do not recognize symlinks to files as files.
+---Return true if the path points at a symbolic link.
 ---@param path string
----@param opts? {preserve_symlinks: boolean?}
----@return OrgPromise<boolean> result
-function M.is_file(path, opts)
-  return M.stat_or_nil(path, opts):next(function(stat)
-    return stat and stat.type == 'file'
-  end)
-end
-
----Resolve to true if the path points at a directory.
----* `preserve_symlinks`: if true, do not recognize symlinks to directories as
----  directories.
----@param path string
----@param opts? {preserve_symlinks: boolean?}
----@return OrgPromise<boolean> result
-function M.is_dir(path, opts)
-  return M.stat_or_nil(path, opts):next(function(stat)
-    return stat and stat.type == 'directory'
-  end)
-end
-
----Resolve to true if the path points at a directory with zero files in it.
----* `preserve_symlinks`: if true, do not recognize symlinks to directories as
----  directories.
----@param path string
----@param opts? {preserve_symlinks: boolean?}
----@return OrgPromise<boolean> result
-function M.is_empty_dir(path, opts)
-  return M.is_dir(path, opts):next(function(is_dir)
-    if is_dir then
-      for _, _ in M.iterdir(path) do
-        return false
-      end
-      return true
-    end
-    return false
-  end)
-end
-
----Resolve to true if the path points at a symbolic link.
----@param path string
----@return OrgPromise<boolean> result
+---@return boolean is_symlink
 function M.is_symlink(path)
-  return M.stat_or_nil(path, { preserve_symlinks = true }):next(function(stat)
-    return stat and stat.type == 'link'
-  end)
+  local link, errmsg, err = uv.fs_readlink(path)
+  assert(err == 'EINVAL' or err == 'ENOENT', errmsg)
+  return link and true or false
 end
 
 ---Helper function to `copy_directory` and `remove_directory`.
@@ -316,10 +209,8 @@ end
 ---@param path string
 ---@return table<string, string[]> map_types_to_names_lists
 local function groupdir(path)
-  local node, err = uv.fs_scandir(path)
-  if not node then
-    error(err)
-  end
+  ---@todo is scandir or readdir the faster option?
+  local node = assert(uv.fs_scandir(path))
   local res = {}
   while true do
     local name, ftype = uv.fs_scandir_next(node)
@@ -357,52 +248,44 @@ end
 function M.copy_symlink(path, new_path, flags)
   local keep_times = flags and flags.keep_times or false
   local exist_ok = flags and flags.exist_ok or false
-  return M.readlink(path):next(function(target)
-    return M.is_dir(target):next(function(dir)
-      return M.symlink(target, new_path, { exist_ok = exist_ok, dir = dir, junction = true })
-    end)
-    ---@param created boolean
-  end):next(function(created)
-    if not keep_times then
-      return created
-    end
-    return M.stat(path):next(function(stat)
-      local atime = to_epoch(stat.atime)
-      local mtime = to_epoch(stat.mtime)
-      return Promise.new(function(resolve, reject)
-        uv.fs_lutime(new_path, atime, mtime, function(err)
-          if err then
-            reject(err)
-          else
-            resolve(created)
-          end
+  local target = assert(uv.fs_readlink(path))
+  return M.symlink(target, new_path, { exist_ok = exist_ok, dir = M.is_dir(target), junction = true })
+      :next(function(created)
+        if not keep_times then
+          return created
+        end
+        local stat = assert(uv.fs_stat(path))
+        local atime = to_epoch(stat.atime)
+        local mtime = to_epoch(stat.mtime)
+        return Promise.new(function(resolve, reject)
+          uv.fs_lutime(new_path, atime, mtime, function(err)
+            if err then
+              reject(err)
+            else
+              resolve(created)
+            end
+          end)
         end)
       end)
-    end)
-  end)
 end
 
 ---Copy permission bits and (potentially) timestamps from one file to another.
 ---@param path string
 ---@param new_path string
 ---@param keep_times boolean if true, copy access and modification timestamps
----@return OrgPromise<nil>
+---@return nil
 local function copy_stats(path, new_path, keep_times)
-  return M.stat(path):next(function(stat)
-    uv.fs_chmod(new_path, stat.mode)
-    if not keep_times then
-      return nil
-    end
-    local atime = to_epoch(stat.atime)
-    local mtime = to_epoch(stat.mtime)
-    uv.fs_utime(new_path, atime, mtime)
-    return M.is_symlink(new_path):next(function(is_new_path_symlink)
-      if is_new_path_symlink then
-        uv.fs_lutime(new_path, atime, mtime)
-      end
-      return nil
-    end)
-  end)
+  local stat, errmsg, err = assert(uv.fs_stat(path))
+  assert(uv.fs_chmod(new_path, stat.mode))
+  if not keep_times then
+    return
+  end
+  local atime = to_epoch(stat.atime)
+  local mtime = to_epoch(stat.mtime)
+  assert(uv.fs_utime(new_path, atime, mtime))
+  if M.is_symlink(new_path) then
+    assert(uv.fs_lutime(new_path, atime, mtime))
+  end
 end
 
 ---The meat of `copy_directory`, without `create_symlink` handling.
@@ -418,7 +301,7 @@ local function copy_directory_impl(path, new_path, opts)
         local items = groupdir(path)
 
         local copy_files = Promise.map(function(name)
-          M.copy_file(
+          return M.copy_file(
             vim.fs.joinpath(path, name),
             vim.fs.joinpath(new_path, name),
             { excl = false, ficlone = true, ficlone_force = false }
@@ -434,19 +317,13 @@ local function copy_directory_impl(path, new_path, opts)
         end, items.link or {}, 1)
 
         local copy_dirs = Promise.map(function(name)
-          M.copy_directory(
-            vim.fs.joinpath(path, name),
-            vim.fs.joinpath(new_path, name),
-            opts
-          )
+          return M.copy_directory(vim.fs.joinpath(path, name), vim.fs.joinpath(new_path, name), opts)
         end, items.directory or {}, 1)
 
         return Promise.all({ copy_files, copy_links, copy_dirs })
       end)
       :next(function()
-        return copy_stats(path, new_path, keep_times)
-      end)
-      :next(function()
+        copy_stats(path, new_path, keep_times)
         return true
       end)
 end
@@ -463,32 +340,18 @@ end
 ---@return OrgPromise<true> success
 function M.copy_directory(path, new_path, opts)
   local create_symlink = opts and opts.create_symlink or false
-  return M.readlink(path):next(function(path_symlink_target)
-    if not create_symlink then
-      return false
+  local path_symlink_target, errmsg, err = uv.fs_readlink(path)
+  if path_symlink_target and create_symlink then
+    if M.is_dir(new_path) then
+      new_path = vim.fs.joinpath(new_path, vim.fs.basename(path))
     end
     -- Source directory is a symbolic link, copy only the link.
-    return M.is_dir(new_path):next(function(new_path_is_dir)
-      if new_path_is_dir then
-        new_path = vim.fs.joinpath(new_path, vim.fs.basename(path))
-      end
-      return M.symlink(path_symlink_target, new_path, {
-        dir = true, junction = true, exist_ok = true })
-    end):next(function() return true end)
-    ---@diagnostic disable-next-line
-  end, function(err)
-    -- readlink failed; either path isn't a symbolic link or another error
-    -- occurred.
-    if vim.startswith(err, 'EINVAL:') then
-      return Promise.resolve(false)
-    end
-    return Promise.reject(err)
-  end):next(function(handled)
-    -- Source directory is not a symbolic link; create destination
-    -- directory, then copy recursively.
-    if handled then return true end
-    return copy_directory_impl(path, new_path, opts)
-  end)
+    return M.symlink(path_symlink_target, new_path, { dir = true, junction = true, exist_ok = true }):next(function()
+      return true
+    end)
+  end
+  assert(err == 'EINVAL', errmsg)
+  return copy_directory_impl(path, new_path, opts)
 end
 
 ---Delete a directory, potentially recursively.
@@ -496,39 +359,33 @@ end
 ---  The default is to only delete `path` if its an empty directory.
 ---@param path string
 ---@param opts? {recursive: boolean?}
----@return OrgPromise<boolean> success
+---@return OrgPromise<true> success
 function M.remove_directory(path, opts)
-  opts = opts or {}
-  local recursive = opts.recursive or false
-  local prev_jobs = {}
-  if recursive then
+  ---@type OrgPromise[]|nil
+  local prior_jobs
+  if opts and opts.recursive then
     local items = groupdir(path)
-    local subdirs = items.directory or {}
-    items.directory = nil
     local rm_dirs = Promise.map(function(subdir)
       return M.remove_directory(vim.fs.joinpath(path, subdir), opts)
-    end, subdirs, 1)
-    local rest = {}
-    for _, names in pairs(items) do
-      vim.list_extend(rest, names)
-    end
+    end, items.directory or {}, 1)
+    items.directory = nil
+    local rest = vim.tbl_flatten(vim.tbl_values(items))
     local rm_files = Promise.map(function(name)
       return M.unlink(vim.fs.joinpath(path, name))
     end, rest, 1)
-    prev_jobs = { rm_dirs, rm_files }
+    prior_jobs = { rm_dirs, rm_files }
   end
-  return Promise.all(prev_jobs)
-      :next(function()
-        return Promise.new(function(resolve, reject)
-          uv.fs_rmdir(path, function(err, success)
-            if err then
-              reject(err)
-            else
-              resolve(success)
-            end
-          end)
-        end)
+  return Promise.all(prior_jobs or {}):next(function()
+    return Promise.new(function(resolve, reject)
+      uv.fs_rmdir(path, function(err, success)
+        if success then
+          resolve(success)
+        else
+          reject(err)
+        end
       end)
+    end)
+  end)
 end
 
 --[[
@@ -562,13 +419,12 @@ end
 ---@param url string
 ---@param dest string
 ---@param opts? {exist_ok: boolean?}
----@return OrgPromise<boolean> success
+---@return OrgPromise<true> success
 function M.download_file(url, dest, opts)
   opts = opts or {}
   local exist_ok = opts.exist_ok or false
   return netrw_read(url):next(function(source)
-    return M.copy_file(source, dest,
-      { excl = not exist_ok, ficlone = true, ficlone_force = false })
+    return M.copy_file(source, dest, { excl = not exist_ok, ficlone = true, ficlone_force = false })
   end)
 end
